pyproject.toml
[tool.poetry]
name = "pocket-playground"
version = "0.1.0"
description = ""
authors = ["kumarstack55 <kumarstack55@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pocket = "^0.3.6"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
